+++ ansible_rulebook/rule_set_runner.py
@@ -247,9 +247,14 @@ class RuleSetRunner:
                     self.active_actions.add(task)
                     task.add_done_callback(self._handle_action_completion)
                 else:
-                    self._run_action(
+                    task = self._run_action(
                         action_item.actions[0], action_item, rule_run_at
                     )
+                    if (
+                        self.parsed_args
+                        and self.parsed_args.execution_strategy == "sequential"
+                    ):
+                        await task
         except asyncio.CancelledError:
             logger.debug(
                 "Action Plan Task Cancelled for ruleset %s", self.name