+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Added
 - Sending heartbeat to the server with the session stats
+- Added command line option --execution-strategy
 
 ### Fixed